apps/mcp-server/pyproject.toml
[project]
name = "skillhub-mcp-server"
version = "1.0.0"
description = "SkillHub MCP server"
requires-python = ">=3.11"
dependencies = [
  "mcp",
  "httpx",
  "pyjwt",
  "pydantic-settings",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp-proto-grpc",
  "opentelemetry-instrumentation-httpx",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "mypy",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
strict = true
python_version = "3.11"
